+++ src/components/Page Header/index.jsx
@@ -1,14 +1,20 @@
 import React from 'react';
+import Proptypes from 'prop-types';
 
 const PageHeader = props => (
-  <section>
+  <section className="pt-4">
     <div className="hero-body">
       <div className="container">
         <h1 className="title">{props.title}</h1>
-        <h2 className="subtitle">{props.children}</h2>
+        <h3 className="subtitle">{props.children}</h3>
       </div>
     </div>
   </section>
 );
+
+PageHeader.propTypes = {
+  title: Proptypes.string.isRequired,
+  children: Proptypes.object.isRequired
+};
 
 export default PageHeader;