+++ src/scenes/Home/components/Mission Section/index.jsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import PageSection from '../../../../components/Page Section';
+
+class MissionSection extends React.Component {
+  render() {
+    return (
+      <PageSection height="50">
+        <div className="row justify-content-center no-gutters">
+          <div className="col-8 col-md-5">
+            <h1 >Mission</h1>
+            <p>
+              Lorem ipsum, dolor sit amet consectetur adipisicing elit.
+              Veritatis reiciendis vitae corrupti laboriosam doloribus hic cum
+              et harum voluptates, beatae iste aliquam earum amet consequatur?
+              Perferendis enim quod facilis dolorum!
+            </p>
+          </div>
+        </div>
+      </PageSection>
+    );
+  }
+}
+
+export default MissionSection;