+++ src/scenes/Home/index.jsx
@@ -1,15 +1,22 @@
 import React from 'react';
 
-import Blog from '../Blog';
+import VideoHeader from '../../components/Video Header';
+import WorkSection from './components/Work Section';
+import TeamSection from './components/Team Section';
+import MissionSection from './components/Mission Section';
+import DonateSection from '../Donate Section';
 
 class Home extends React.Component {
   render() {
     return (
-      <div>
-        <h1>The Lilac Foundation</h1>
-      </div>
+      <React.Fragment>
+        <VideoHeader />
+        <MissionSection />
+        <WorkSection />
+        <TeamSection />
+        <DonateSection />
+      </React.Fragment>
     );
   }
 }
-
 export default Home;