+++ changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.0.2
+
+### Added
+
+#### 2018-10-21
+
+- Implemented home page layout with styling
+  - Added Mission section
+  - Added Team section
+  - Added Work section
+
 ## 0.0.1
 
 ### Added
@@ -19,11 +30,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added colours [SASS](https://sass-lang.com/) module
 
 - Restructured app and routing structure
+- Added env file for imports
 
 - Added [React Burger Menu](http://negomi.github.io/react-burger-menu/) for mobile navigation
 - Implemented responsive navbar
-
-- Added env file for imports
 
 - Added footer with styling